+++ src/main/java/edu/clemson/cs/r2jt/proving/absyn/PAlternatives.java
@@ -23,8 +23,8 @@ import java.util.Set;
 import edu.clemson.cs.r2jt.absyn.AltItemExp;
 import edu.clemson.cs.r2jt.absyn.AlternativeExp;
 import edu.clemson.cs.r2jt.typeandpopulate.MTType;
-import edu.clemson.cs.r2jt.proving.immutableadts.ArrayBackedImmutableList;
-import edu.clemson.cs.r2jt.proving.immutableadts.ImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ArrayBackedImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ImmutableList;
 import edu.clemson.cs.r2jt.utilities.Mapping;
 import edu.clemson.cs.r2jt.utilities.RCollections;
