+++ src/main/java/edu/clemson/cs/r2jt/proving/absyn/PSymbol.java
@@ -22,10 +22,9 @@ import java.util.Map;
 import java.util.Set;
 
 import edu.clemson.cs.r2jt.typeandpopulate.MTType;
-import edu.clemson.cs.r2jt.proving.immutableadts.ArrayBackedImmutableList;
-import edu.clemson.cs.r2jt.proving.immutableadts.ImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ArrayBackedImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ImmutableList;
 import edu.clemson.cs.r2jt.typeandpopulate.MTFunction;
-import java.util.Collections;
 
 /**
  * <p>A <code>PSymbol</code> represents a reference to a named element such as
