+++ src/main/java/edu/clemson/cs/r2jt/proving2/AutomatedProver.java
@@ -13,7 +13,7 @@
 package edu.clemson.cs.r2jt.proving2;
 
 import edu.clemson.cs.r2jt.ResolveCompiler;
-import edu.clemson.cs.r2jt.proving.immutableadts.ImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ImmutableList;
 import edu.clemson.cs.r2jt.proving2.applications.Application;
 import edu.clemson.cs.r2jt.proving2.automators.AntecedentDeveloper;
 import edu.clemson.cs.r2jt.proving2.automators.AntecedentMinimizer;
