+++ src/main/java/edu/clemson/cs/r2jt/proving2/automators/AntecedentMinimizer.java
@@ -12,7 +12,7 @@
  */
 package edu.clemson.cs.r2jt.proving2.automators;
 
-import edu.clemson.cs.r2jt.proving.immutableadts.ImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ImmutableList;
 import edu.clemson.cs.r2jt.proving2.model.PerVCProverModel;
 import edu.clemson.cs.r2jt.proving2.model.Theorem;
 import edu.clemson.cs.r2jt.proving2.transformations.ExpandAntecedentBySubstitution;
