+++ src/main/java/edu/clemson/cs/r2jt/proving2/iterators/ChainingIterator.java
@@ -10,7 +10,7 @@
  * This file is subject to the terms and conditions defined in
  * file 'LICENSE.txt', which is part of this source code package.
  */
-package edu.clemson.cs.r2jt.proving;
+package edu.clemson.cs.r2jt.proving2.iterators;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
