+++ src/main/java/edu/clemson/cs/r2jt/typeandpopulate/searchers/OperationProfileSearcher.java
@@ -14,7 +14,7 @@ package edu.clemson.cs.r2jt.typeandpopulate.searchers;
 
 import edu.clemson.cs.r2jt.data.Location;
 import edu.clemson.cs.r2jt.data.PosSymbol;
-import edu.clemson.cs.r2jt.proving.immutableadts.ImmutableList;
+import edu.clemson.cs.r2jt.proving2.immutableadts.ImmutableList;
 import edu.clemson.cs.r2jt.typeandpopulate.DuplicateSymbolException;
 import edu.clemson.cs.r2jt.typeandpopulate.SymbolTable;
 import edu.clemson.cs.r2jt.typeandpopulate.entry.OperationProfileEntry;
