+++ src/main/java/edu/clemson/cs/r2jt/vcgeneration/vcs/VerificationCondition.java
@@ -17,7 +17,7 @@ package edu.clemson.cs.r2jt.vcgeneration.vcs;
  */
 import edu.clemson.cs.r2jt.absyn.Exp;
 import edu.clemson.cs.r2jt.collections.List;
-import edu.clemson.cs.r2jt.proving.Conjuncts;
+import edu.clemson.cs.r2jt.proving2.Conjuncts;
 
 public class VerificationCondition {
